feat(ws-client): reject pending requests when closing (#178)
* feat(ws-client): handle socket close

* Test

* Add delay fn to zome

* Write test for connection closing

* lint

* remove unneeded async from client

* delete unused function from test

* enable all tests

* add delay between zome calls

* handle varying response ids for pending requests

* ci(test): pin nix install to bug free version

---------

+++ test/e2e/util.ts
@@ -1,10 +1,10 @@
+import assert from "node:assert/strict";
 import { spawn } from "node:child_process";
 import { Test } from "tape";
 import { AdminWebsocket } from "../../src/api/admin/websocket.js";
 import { AppWebsocket } from "../../src/api/app/websocket.js";
-import { CellId, InstalledAppId } from "../../src/types.js";
-import assert from "node:assert/strict";
 import { AppAgentWebsocket, CellType } from "../../src/index.js";
+import { CellId, InstalledAppId } from "../../src/types.js";
 
 export const FIXTURE_PATH = "./test/e2e/fixture";
 
@@ -81,7 +81,7 @@ export const withConductor =
       console.error("Test caught exception: ", e);
       throw e;
     } finally {
-      if (conductorProcess.pid) {
+      if (conductorProcess.pid && !conductorProcess.killed) {
         process.kill(-conductorProcess.pid);
       }
       await cleanSandboxConductors();
